rest-api/src/main/java/com/company/webshop/service/AccountServiceImplementation.java
package com.company.webshop.service;

import com.company.webshop.domain.Account;
import com.company.webshop.repository.AccountRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class AccountServiceImplementation implements AccountService {

    @Autowired
    private AccountRepository accountRepository;

    @Override
    public Account getAccountById(Long id) {
        return accountRepository.findById(id).get();
    }

    @Override
    public Account createAccount(Account account) {
        return accountRepository.save(account);
    }

    @Override
    public void deleteAllAccounts() {
        accountRepository.deleteAllInBatch();
    }
}
